- Refactor exception handling and improve service layer logic - Add testcases
The commit removes direct dependency on BookNotFoundException and UserNotFoundException. These exceptions are replaced with built-in EntityNotFoundException which simplifies logic. The commit also changes the AuthenticatedUsername annotation to AuthenticatedUserEmail to reflect the fact that it's an email, not a username that is being fetched. Moreover, the commit improves the service layer by implementing a better way of dealing with file paths in services related to the handling of book covers.